% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.tuning}
\alias{select.tuning}
\title{Tuning paramater selection}
\usage{
select.tuning(object, figure = FALSE, criteria = c("BIC", "GIC", "AIC"))
}
\arguments{
\item{object}{Object generated from \code{path.fmrHP}.}

\item{figure}{incidator for showing plot of information criteria.}

\item{criteria}{information criteria for selection of tuning parameter.}
}
\value{
list of parameters of selected model.
}
\description{
Select tuning parameter via AIC, BIC or GIC from objects generated by 
\code{path.fmrHP}.
}
